Fixed scrollbar in dark mode on home page. Added Drive file option when creating new file. Made ID display on view post only if in DEV mode (url origin contains localhost). Added button to search articles of the same category kind. Bug fixes.

+++ src/pages/admin/manageArticles.tsx
@@ -38,7 +38,7 @@ function ArticleRow({value, index}: any) {
     const deleteArticle = () => {
         let file = value.file
         let promises = []
-        if (file) {
+        if (file && value.contentKind !== "drive") {
             let fileDeletionPromise = firebaseStorage.deleteObject(
                 firebaseStorage.ref(
                     firebaseStorage.getStorage(),
